refactor(dev-preview): extract terminal lifecycle controller
- Move terminal spawn/attach/detach/stop/wait control, install spawning, and
  exit handling (ensureSessionTerminal, spawnSessionTerminal, runInstall,
  handleExit, etc.) into DevPreviewTerminalController.ts
- Deps bag (ptyClient, portRegistry, terminalToSession, portWaitAborts,
  isDisposed, recordDiagnostic, recordSessionDiagnostic, updateSession,
  clearCompiling, pollServerReadiness) keeps the import arrow
  orchestrator -> sibling; handleDevPreviewTerminalExit imports
  guardedReinstall and classifyDevPreviewExit directly from their sibling
  modules (one-directional sibling-to-sibling, no cycle)
- DevPreviewSessionService keeps thin wrappers backed by a single
  terminalControllerDeps getter; removes the now-dead classifyExit/
  guardedReinstall wrappers whose only caller (handleExit) was replaced
- Preserves the exact '[DevPreviewSessionService]' log prefixes for
  observable warn() messages
